Move Resources::Base back to Resource
