+++ lib/cm_api/client/users.rb
@@ -6,7 +6,7 @@ module CMAPI
 
       # Gets a list of CDM users
       #
-      # @return [Resources::Base] the list of users
+      # @return [Resource] the list of users
       def users
         get("/users")
       end
@@ -24,7 +24,7 @@ module CMAPI
       # @param name [String] the new user's username
       # @param password [String] the user's password
       # @param roles [Array<String>] the roles that the user belongs to (default: "ROLE_USER")
-      # @return [Resources::Base] the new user
+      # @return [Resource] the new user
       def create_user(name:, password:, roles: [])
         body = { name: name, password: password, roles: Array(roles) }
 
@@ -37,7 +37,7 @@ module CMAPI
       # @param name [String] the user's username
       # @param password [String] the new password (pass `nil` to skip update)
       # @param roles [Array<String>] the new roles for the user (leave empty to skip update)
-      # @return [Resources::Base] the updated user
+      # @return [Resource] the updated user
       def update_user(name:, password: nil, roles: [])
         return user(name: name) if password.blank? && roles.blank?
 
@@ -51,14 +51,14 @@ module CMAPI
       # Deletes the specified user from CDM
       #
       # @param name [String] the user to delete
-      # @return [Resources::Base] the deleted user
+      # @return [Resource] the deleted user
       def delete_user(name:)
         delete("/users/#{name}")
       end
 
       # Looks up current user sessions
       #
-      # @return [Resources::Base] the currently active sessions
+      # @return [Resource] the currently active sessions
       def user_sessions
         get("/users/sessions")
       end
