+++ lib/cm_api/middleware/response.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 module CMAPI
   module Middleware
-    # Response middleware for converting a response body into a {Resources::Base}
+    # Response middleware for converting a response body into a {Resource}
     class Response < Faraday::Response::Middleware
       # @return [Client] the api client that initiated the request
       attr_reader :api_client
@@ -14,10 +14,10 @@ module CMAPI
         @api_client = api_client
       end
 
-      # Parse the response into a {Resources::Base} object(s)
+      # Parse the response into a {Resource} object(s)
       #
       # @param response [String] the response body
-      # @return [Resources::Base, Array<Resources::Base>] the parsed resource(s)
+      # @return [Resource, Array<Resource>] the parsed resource(s)
       #
       # @note when the response contains a single top-level item named "items", each resource will be parsed and an
       # array will be returned.
@@ -29,7 +29,7 @@ module CMAPI
       private
 
       def create_resource(item)
-        item.is_a?(Hash) ? Resources::Base.new(item, api_client: api_client) : item
+        item.is_a?(Hash) ? Resource.new(item, api_client: api_client) : item
       end
 
       def array_response?(json)
